+++ src/pytest_reverse.py
@@ -1,3 +1,6 @@
+import pytest
+
+
 def pytest_addoption(parser):
     group = parser.getgroup("reverse", "Reverse test order")
     group._addoption(
@@ -8,6 +11,9 @@ def pytest_addoption(parser):
     )
 
 
+@pytest.hookimpl(hookwrapper=True, tryfirst=True)
 def pytest_collection_modifyitems(session, config, items):
     if config.getoption("reverse"):
         items[:] = items[::-1]
+
+    yield