+++ HISTORY.rst
@@ -2,6 +2,9 @@
 History
 -------
 
+* Run first to avoid interfering with ``--failed-first`` flag
+  (`docs <https://docs.pytest.org/en/latest/cache.html>`__).
+
 1.0.0 (2020-04-03)
 ------------------